Change propagation/receiving functions
It isn't necessary to exchange the whole blockchain each update. So, I
classified situations into two cases.
First one is that a new block is created, so every node should get the new
block and add it at the end of its own chain. By doing this, existing
blocks don't get affected, so the game goes well.
Second one is that a transaction is updated, so only nodes of a specific
game block have to renew their transaction.